Renamed VLA ray tracing to Cassegrain ray tracing
