+++ src/astrohack/config/cassegrain_ray_tracing.param.json
