+++ src/astrohack/core/cassegrain_ray_tracing.py
